分數計算方式及時間應該無誤，加上上傳sql的功能，有些testdata可能因電腦效能太爛跑很慢然後當掉

+++ 1/1/fib.cpp
@@ -2,13 +2,14 @@
 #include <iostream>
 #include <sstream>
 #include <string>
+#include <math.h>
 using namespace std;
 Fib::Fib()
 {
 
 }
 
-void Fib::fib_last_digit(int *f,int n){
+void Fib::fib_last_digit(int *f,long long int n){
     f[0]=0;
     f[1]=1;
     for(int i=2;i<n;i++){
@@ -16,11 +17,11 @@ void Fib::fib_last_digit(int *f,int n){
     }
 }
 
-int Fib::find_last_digit(int *f,int n){
+int Fib::find_last_digit(int *f,long long int n){
     return f[n];
 }
 
-int Fib::shift(int n){
+int Fib::shift(long long int n){
     int f0=0,f1=1;
     int result=1;
     for(int i=2;i<=n;i++){
@@ -31,20 +32,25 @@ int Fib::shift(int n){
         f0 = f1;
         f1 = result;
     }
+//    long  long int result;
+//    double estim = 0.4472135955 * pow(1.618033988745, n+1);
+//    result =  ( abs(ceil(estim)-estim) < abs(estim-floor(estim)) )? ceil(estim) : floor(estim);
+//    result%=60;
     return  result;
 }
 
 string Fib::solve(string s){
     stringstream ss;
     ss<<s;
-    int n;
+    long long int n;
     string result="";
     int ans;
     //first digit cycle every 60 numbers
     int f[60];
     fib_last_digit(f,60);
-    int temp;   //temp = fib(n)%60
+    long long int temp;   //temp = fib(n)%60
     while(ss>>n){
+//        cout<<n<<endl;
         temp = shift(n);
         ans = find_last_digit(f,temp);
         result.append(to_string(ans)+' ');
